requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.13.3
branca==0.7.2
contextily==1.5.0
earthengine_api==0.1.400
earthengine_api==0.1.410
ee==0.2
Flask==3.1.0
Flask_Caching==2.3.0
Flask_Cors==4.0.0
folium==0.17.0
GDAL==3.8.2
geemap==0.32.0
geojson==3.1.0
geopandas==0.14.1
geopy==2.4.1
googlemaps==4.10.0
ipython==8.12.3
ipython==8.26.0
Jinja2==3.1.5
matplotlib==3.8.2
networkx==3.3
numpy==2.2.2
osmnx==2.0.1
pandas==2.2.3
PyCRS==1.0.2
python-dotenv==1.0.1
Requests==2.32.3
Shapely==2.0.7
